Set serial port to blocking wile sending image
The serial port needs to be in non-blocking mode for normal operation to
use the select() loop. However, if the image being sent is especially
big (e.g. a large operating system), then the write() will eventually
need to block. So clear the non-blocking flag while sending the image,
and then set it once the image is sent.